Unsplash

+++ spec/requests/api/v1/forecast_spec.rb
@@ -35,7 +35,7 @@ describe 'forecast endpoint', :vcr do
     expect(weather[:data][:attributes][:current_weather]).to have_key(:humidity)
     expect(weather[:data][:attributes][:current_weather][:humidity]).to be_a Integer
     expect(weather[:data][:attributes][:current_weather]).to have_key(:uvi)
-    expect(weather[:data][:attributes][:current_weather][:uvi]).to be_a Float
+    expect(weather[:data][:attributes][:current_weather][:uvi]).to be_a Integer
     expect(weather[:data][:attributes][:current_weather]).to have_key(:visibility)
     expect(weather[:data][:attributes][:current_weather][:visibility]).to be_a Integer
     expect(weather[:data][:attributes][:current_weather]).to have_key(:conditions)
